@@ -14,20 +14,22 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.18.0"
 
 dependencies:
   flutter:
     sdk: flutter
+    version: ^0.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  url_launcher: ^5.1.2
+  cupertino_icons: ^1.0.6
+  url_launcher: ^6.2.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+    version: ^0.0.0
 
 
 # For information on the generic Dart part of this file, see the
